+++ src/main/java/com/api/safectech/publicationservice/comment/domain/service/CommentService.java
@@ -12,5 +12,5 @@ public interface CommentService {
     Comment update(Long commentId, Comment request);
     ResponseEntity<?> delete(Long commentId);
     List<Comment> getByPublicationId(Long publicationId);
-    List<Comment> getByUserId(Integer userId);
+    List<Comment> getByUserId(Long userId);
 }
